feat(ai): add getOllamaBaseUrl helper for native API URL resolution
Extracts getOllamaBaseUrl into ollama-url.ts (no DB dependencies) so
the pure function can be tested in isolation without mocking @dpf/db.
ollama.ts re-exports it for backward compatibility.
